+++ DirectionalCompressor/Source/PluginEditor.h
@@ -30,7 +30,7 @@
 #include "../../resources/customComponents/TitleBar.h"
 #include "../../resources/customComponents/LevelMeter.h"
 #include "../../resources/customComponents/SimpleLabel.h"
-#include "../../resources/customComponents/IEMSphere.h"
+#include "../../resources/customComponents/SpherePanner.h"
 
 typedef ReverseSlider::SliderAttachment SliderAttachment;
 typedef AudioProcessorValueTreeState::ComboBoxAttachment ComboBoxAttachment;
@@ -40,7 +40,7 @@ typedef AudioProcessorValueTreeState::ButtonAttachment ButtonAttachment;
 /**
 */
 class DirectionalCompressorAudioProcessorEditor  : public AudioProcessorEditor,
-private Timer, public IEMSphere::IEMSphereListener, public IEMSphere::IEMSphereElement, private Button::Listener
+private Timer, private Button::Listener
 {
 public:
     DirectionalCompressorAudioProcessorEditor (DirectionalCompressorAudioProcessor&, AudioProcessorValueTreeState&);
@@ -49,7 +49,6 @@ public:
     //==============================================================================
     void paint (Graphics&) override;
     void resized() override;
-    void IEMSphereElementChanged (IEMSphere* sphere, IEMSphereElement* element) override;
     void buttonStateChanged (Button* button) override;
     void buttonClicked (Button* button) override {};
     
@@ -59,12 +58,11 @@ private:
     DirectionalCompressorAudioProcessor& processor;
     AudioProcessorValueTreeState& valueTreeState;
     
-    TitleBar<AmbisonicIOWidget<>,NoIOWidget> title;
+    TitleBar<AmbisonicIOWidget<>, NoIOWidget> title;
     Footer footer;
     
-    IEMSphere sphere;
-    IEMSphereElement sphereElem;
-    
+    SpherePanner sphere;
+    SpherePanner::AziumuthElevationParameterElement sphereElem;
     
     int maxPossibleOrder = -1;
     ScopedPointer<ComboBoxAttachment> cbNormalizationAtachement;
@@ -80,7 +78,7 @@ private:
     ToggleButton tbC1;
     ToggleButton tbC2;
     
-    ReverseSlider slPreGain, slYaw, slPitch, slWidth;
+    ReverseSlider slPreGain, slAzimuth, slElevation, slWidth;
     ReverseSlider slC1Threshold, slC1Knee, slC1Ratio, slC1Attack, slC1Release, slC1Makeup;
     ReverseSlider slC2Threshold, slC2Knee, slC2Ratio, slC2Attack, slC2Release, slC2Makeup;
     
@@ -88,7 +86,7 @@ private:
     ComboBox cbC2Driving, cbC2Apply;
     ComboBox cbListen;
     
-    ScopedPointer<SliderAttachment> slPreGainAttachment, slYawAttachment, slPitchAttachment, slWidthAttachment;
+    ScopedPointer<SliderAttachment> slPreGainAttachment, slAzimuthAttachment, slElevationAttachment, slWidthAttachment;
     ScopedPointer<SliderAttachment> slC1ThresholdAttachment, slC1KneeAttachment, slC1RatioAttachment;
     ScopedPointer<SliderAttachment> slC1AttackAttachment, slC1ReleaseAttachment, slC1MakeupAttachment;
     ScopedPointer<SliderAttachment> slC2ThresholdAttachment, slC2KneeAttachment, slC2RatioAttachment;
@@ -103,7 +101,7 @@ private:
     LevelMeter dbC1GRmeter, dbC1RMSmeter;
     LevelMeter dbC2GRmeter, dbC2RMSmeter;
     
-    SimpleLabel lbPreGain, lbYaw, lbPitch, lbWidth;
+    SimpleLabel lbPreGain, lbAzimuth, lbElevation, lbWidth;
     SimpleLabel lbC1Threshold, lbC1Knee, lbC1Ratio, lbC1Attack, lbC1Release, lbC1Makeup;
     SimpleLabel lbC2Threshold, lbC2Knee, lbC2Ratio, lbC2Attack, lbC2Release, lbC2Makeup;